speed test added
